Stop creating empty branch if no changes

+++ scripts/git-commit.sh
@@ -16,6 +16,13 @@ commit_message="Workflow: created files in ${1}"
 git checkout -b "$branch"
 git add "$1"
 git commit -m "$commit_message"
+
+commit_success=$?
+if [ $commit_success -ne 0 ]; then
+  echo "Nothing to commit"
+  exit 0
+fi
+
 git remote rm origin || true
 git remote add origin "https://${TOKEN}@github.com/${GITHUB_REPOSITORY}.git"
 git push -u origin "$branch"
